Checkerをコンポーネント化
