+++ src/app/colors/contrasts/_components/check-contrast/check-contrast.stories.tsx
@@ -1,0 +1,13 @@
+import type { Meta, StoryObj } from '@storybook/react';
+import { CheckContrast } from './check-contrast';
+
+const meta: Meta<typeof CheckContrast> = {
+  title: 'app/colors/contrasts/check-contrast',
+  component: CheckContrast,
+  tags: ['autodocs'],
+};
+
+export default meta;
+type Story = StoryObj<typeof CheckContrast>;
+
+export const Default: Story = {};
