+++ src/app/colors/contrasts/_components/check-contrast/index.ts
@@ -1,0 +1,1 @@
+export * from './check-contrast';
